+++ audio/src/fetch/mod.rs
@@ -301,12 +301,12 @@ impl StreamLoaderController {
         self.send_stream_loader_command(StreamLoaderCommand::Close);
     }
 
-    pub fn from_local_file(file_size: u64) -> Result<Self, Error> {
-        Ok(Self {
+    pub fn from_local_file(file_size: u64) -> Self {
+        Self {
             channel_tx: None,
             stream_shared: None,
             file_size: file_size as usize,
-        })
+        }
     }
 }